Hide web only elements when printing from plone
When printing from plone certain elements are visible on the printed page,
especially the access links. With the default plone print style
(plonePrint.css) all links get expanded to full URLs which results in a lot of
useless junk at the top of the screen. This patch adds a few plone related CSS
classes to certain sections of the page which in turn ensures they aren't
visible when printed.

This has no effect on browsers which aren't aware of the print style sheet. The
access keys still work fine in FireFox 1.0.