1303 Illegal utf-8 handling in Utils.py summary - credits to hideo at yokohama - http://my.opera.com/hideo_at_yokohama
Credits go to Hideo at Yokohama! This patch had us worried about performance, 
since taking a summary of a long wiki page might be slow. Some preliminary testing
showed that the performance price is rather low, but for sites with non-ascii
content the patch is really worth it. Not only will it stop producing illegal
utf-8, but without the patch the lenght of the summary will never be right.