siblingsAsList can include current page + sorting optional
This is useful mainly for "nautica" style wikipage.pt. In those
one often wants to make a list of pages "on the same level" as
the current page. That list usually includes the current page
and is not sorted alphabetically, but instead left in the order
defined in the /backlinks page.